backend/requirements.txt
fastapi
uvicorn
python-dotenv
confluent-kafka
openai==1.12.0   # Azure OpenAI compatible client
